all post routes are confirmed working except for comment and like routes - need models for those
