+++ client/src/graphql/mutations.js
@@ -65,71 +65,6 @@ export const DELETE_POST = gql`
   }
 `;
 
-export const GET_POST = gql`
-  query getPost($postId: ID!) {
-    getPost(postId: $postId) {
-      id
-      imageUrl
-      location
-      title
-      content
-      createdAt
-      userId
-      postLikes {
-        id
-        userId
-        createdAt
-      }
-      postLikeCount
-      comments {
-        id
-        content
-        createdAt
-        userId
-        commentLikes {
-          id
-          userId
-          createdAt
-        }
-        commentLikeCount
-      }
-    }
-  }
-`;
-
-export const GET_POSTS = gql`
-  {
-    getPosts {
-      id
-      imageUrl
-      location
-      title
-      content
-      createdAt
-      userId
-      postLikes {
-        id
-        userId
-        createdAt
-      }
-      postLikeCount
-      commentCount
-      comments {
-        id
-        content
-        createdAt
-        userId
-        commentLikes {
-          id
-          userId
-          createdAt
-        }
-        commentLikeCount
-      }
-    }
-  }
-`;
-
 export const LIKE_POST = gql`
   mutation likePost($postId: ID!) {
     likePost(postId: $postId) {
